+++ liesel/model/nodes.py
@@ -12,7 +12,7 @@ from collections.abc import Callable, Hashable, Iterable
 from functools import wraps
 from itertools import chain
 from types import MappingProxyType
-from typing import TYPE_CHECKING, Any, NamedTuple, TypeVar, Union
+from typing import TYPE_CHECKING, Any, NamedTuple, TypeGuard, TypeVar, Union
 
 import tensorflow_probability.substrates.jax.bijectors as jb
 import tensorflow_probability.substrates.jax.distributions as jd
@@ -1041,6 +1041,10 @@ class NoDist(Dist):
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 
+def is_bijector_class(obj) -> TypeGuard[type[Any]]:
+    return isinstance(obj, type) and issubclass(obj, jb.Bijector)
+
+
 class Var:
     """
     A variable in a statistical model, typically with a probability distribution.
@@ -1074,6 +1078,9 @@ class Var:
 
     See Also
     --------
+    :meth:`.Var.transform` : Transforms a variable by adding a new transformed
+        variable as an input. This is useful for variables that are constrained to a
+        certain domain, e.g. positive values.
     .obs : Helper function to declare a variable as an observed quantity.
     .param : Helper function to declare a variable as a model parameter.
     .Calc :
@@ -1252,6 +1259,170 @@ class Var:
                 visited.append(node)
 
         return _unique_tuple(_vars)
+
+    def transform(
+        self,
+        bijector: type[jb.Bijector] | jb.Bijector | None = None,
+        *bijector_args,
+        **bijector_kwargs,
+    ) -> Var:
+        """
+        Transforms the variable, making it a function of a new variable.
+
+        Creates a new variable on the unconstrained space ``R**n`` with the appropriate
+        transformed distribution, turning the original variable into a weak variable
+        without an associated distribution. The transformation is performed using
+        TFP's bijector classes.
+
+
+        Parameters
+        ----------
+        bijector
+            The bijector used to map the new transformed variable to this variable \
+            (forward transformation). If ``None``, the experimental default event \
+            space bijector (see tensorflow probability documentation) is used. \
+            If a bijector class is \
+            passed, it is instantiated with the arguments ``bijector_args`` and \
+            ``bijector_kwargs``. If a bijector instance is passed, it is used \
+            directly.
+        bijector_args
+            The arguments passed on to the init function of the bijector.
+        bijector_kwargs
+            The keyword arguments passed on to the init function of the bijector.
+
+        Returns
+        -------
+        The new transformed variable which acts as an input to this variable.
+
+        Raises
+        ------
+        RuntimeError
+            If the variable is weak or if the variable has no distribution.
+        ValueError
+            If the argument ``bijector`` is ``None``, but the distribution does
+            not have a default event space bijector.
+
+        Notes
+        -----
+
+        This is a simplified pseudo-code illustration of what this method does:
+
+        .. code-block:: python
+
+        import tensorflow_probability.substrates.jax.bijectors as tfb
+        import tensorflow_probability.substrates.jax.distributions as tfd
+
+        def transform(original_var: lsl.Var, bijector: tfb.Bijector):
+            original_dist = original_var.dist_node.distribution
+            dist_inputs = original_var.dist_node.inputs
+
+            # transform the distribution
+            new_dist = tfd.TransformedDistribution(original_dist, tfb.Invert(bijector))
+
+            # transform initial value
+            new_value = bijector.inverse(original_var.value)
+
+            # initialise the new variable
+            new_var = lsl.Var(
+                new_value,
+                lsl.Dist(new_dist, *dist_inputs),
+                name=f"{original_var.name}_transformed"
+            )
+            new_var.parameter = original_var.parameter
+
+            # define the original variable as a function of the new variable
+            original_var.value_node = lsl.Calc(bijector.forward, new_var)
+            original_var.parameter = False
+
+            # return the new variable
+            return new_var
+
+        The value of the attribute :attr:`~liesel.model.nodes.Var.parameter` is
+        transferred to the transformed variable and set to ``False`` on the original
+        variable. The attributes :attr:`~liesel.model.nodes.Var.observed` and
+        :attr:`~liesel.model.nodes.Var.role` have the default values for
+        the transformed variable and remain unchanged on the original variable.
+
+        Examples
+        --------
+
+        >>> import tensorflow_probability.substrates.jax.distributions as tfd
+        >>> import tensorflow_probability.substrates.jax.bijectors as tfb
+
+        Assume we have a variable ``scale`` that is constrained to be positive, and
+        we want to include the log-transformation of this variable in the model.
+        We first set up the parameter var with its distribution:
+
+        >>> prior = lsl.Dist(tfd.HalfCauchy, loc=0.0, scale=25.0)
+        >>> scale = lsl.param(1.0, prior, name="scale")
+
+        The we transform the variable to the log-scale:
+
+        >>> log_scale = scale.transform(tfb.Exp())
+        >>> log_scale
+        Var(name="scale_transformed")
+
+        Now the ``log_scale`` has a log probability, and the ``scale`` variable does
+        not:
+
+        >>> log_scale.update().log_prob
+        Array(-3.6720574, dtype=float32)
+
+        >>> scale.update().log_prob
+        0.0
+        """
+        if self.weak:
+            raise RuntimeError(f"{repr(self)} is weak")
+
+        if self.dist_node is None:  # type: ignore
+            raise RuntimeError(f"{repr(self)} has no distribution")
+
+        if is_bijector_class(bijector) and not (bijector_args or bijector_kwargs):
+            raise ValueError(
+                "You passed a bijector class instead of an instance, but did not "
+                "provide any arguments for the bijector. You should either provide "
+                "arguments or pass an instance of the bijector class instead."
+            )
+
+        # avoid infinite recursion
+        self.auto_transform = False
+
+        # use default event space bijector if bijector is None
+        use_default_bijector = bijector is None
+        if use_default_bijector:
+            dist_inst = self.dist_node.init_dist()
+            default_bijector = dist_inst.experimental_default_event_space_bijector
+
+        if use_default_bijector and default_bijector is None:
+            raise RuntimeError(
+                f"{self} has distribution without default event space bijector "
+                "and no bijector was given"
+            )
+
+        if is_bijector_class(bijector) or use_default_bijector:
+            tvar = _transform_var_with_bijector_class(
+                self, bijector, *bijector_args, **bijector_kwargs
+            )
+        elif isinstance(bijector, jb.Bijector):
+            if bijector_args or bijector_kwargs:
+                raise RuntimeError(
+                    "You passed a bijector instance and "
+                    "nonempty bijector arguments. You should either initialise your "
+                    "bijector directly with the arguments, or pass a bijector class "
+                    "instead. The first option is preferred, if the bijector arguments"
+                    "are constant."
+                )
+            tvar = _transform_var_with_bijector_instance(self, bijector)
+        else:
+            raise TypeError(
+                f"Argument {bijector=} is of invalid type {type(bijector)}."
+            )
+
+        tvar.parameter = self.parameter  # type: ignore
+        self.parameter = False
+        self.dist_node = None
+
+        return tvar
 
     @in_model_method
     def all_output_nodes(self) -> tuple[Node, ...]:
@@ -1537,6 +1708,98 @@ class Var:
 
     def __repr__(self) -> str:
         return f'{type(self).__name__}(name="{self.name}")'
+
+
+def _transform_var_with_bijector_instance(var: Var, bijector_inst: jb.Bijector) -> Var:
+    if var.dist_node is None:  # type: ignore
+        raise RuntimeError(f"{var} has no distribution")
+    InputDist = var.dist_node.distribution
+    inputs = var.dist_node.inputs
+    kwinputs = var.dist_node.kwinputs
+
+    bijector_inv = jb.Invert(bijector_inst)
+
+    def transform_dist(*args, **kwargs):
+        return jd.TransformedDistribution(InputDist(*args, **kwargs), bijector_inv)
+
+    transformed_dist = Dist(
+        transform_dist,
+        *inputs,
+        **kwinputs,
+        _name="",
+        _needs_seed=var.dist_node.needs_seed,
+    )
+
+    transformed_dist.per_obs = var.dist_node.per_obs
+
+    transformed_var = Var(
+        bijector_inv.forward(var.value),
+        transformed_dist,
+        name=f"{var.name}_transformed",
+    )
+
+    var.value_node = Calc(bijector_inst.forward, transformed_var)
+    return transformed_var
+
+
+def _transform_var_with_bijector_class(
+    var: Var, bijector_cls: type[jb.Bijector] | None, *args, **kwargs
+) -> Var:
+    if var.dist_node is None:  # type: ignore
+        raise RuntimeError(f"{var} has no distribution")
+    InputDist = var.dist_node.distribution
+
+    dist_inputs = InputGroup(
+        *var.dist_node.inputs,
+        **var.dist_node.kwinputs,  # type: ignore
+    )
+
+    bijector_inputs = InputGroup(*args, **kwargs)
+
+    # define distribution "class" for the transformed var
+    def transform_dist(dist_args: ArgGroup, bijector_args: ArgGroup):
+        tfp_dist = InputDist(*dist_args.args, **dist_args.kwargs)
+        bjargs, bjkwargs = bijector_args.args, bijector_args.kwargs
+
+        if bijector_cls is None:
+            default_bijector_cls = tfp_dist.experimental_default_event_space_bijector
+            bijector_inst = default_bijector_cls(*bjargs, **bjkwargs)
+        else:
+            bijector_inst = bijector_cls(*bjargs, **bjkwargs)
+
+        bijector_inv = jb.Invert(bijector_inst)
+
+        transformed_dist = jd.TransformedDistribution(
+            tfp_dist, bijector_inv, validate_args=tfp_dist.validate_args
+        )
+
+        return transformed_dist
+
+    dist_node_transformed = Dist(
+        transform_dist,
+        dist_inputs,
+        bijector_inputs,
+        _name="",
+        _needs_seed=var.dist_node.needs_seed,
+    )
+
+    dist_node_transformed.per_obs = var.dist_node.per_obs
+
+    bijector_inv = dist_node_transformed.init_dist().bijector
+
+    transformed_var = Var(
+        bijector_inv.forward(var.value),
+        dist_node_transformed,
+        name=f"{var.name}_transformed",
+    )
+
+    def bijector_fn(value, dist_inputs, bijector_inputs):
+        bijector = transform_dist(dist_inputs, bijector_inputs).bijector
+        return bijector.inverse(value)
+
+    var.value_node = Calc(bijector_fn, transformed_var, dist_inputs, bijector_inputs)
+
+    return transformed_var
 
 
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~